improved config check
