+++ aws_extension/cloud_api_manager/api_manager.py
@@ -6,6 +6,7 @@ import requests
 import utils
 from aws_extension.auth_service.simple_cloud_auth import cloud_auth_manager, Admin_Role
 from aws_extension.cloud_api_manager.api import api
+from utils import has_config
 
 logger = logging.getLogger(__name__)
 logger.setLevel(utils.LOGGING_LEVEL)
@@ -146,6 +147,9 @@ class CloudApiManager:
             if self.auth_manger.enableAuth and not user_token:
                 return []
 
+            if not has_config():
+                return []
+
             response = requests.get(f'{self.auth_manger.api_url}endpoints',
                                     params={
                                         'username': username,
@@ -185,6 +189,9 @@ class CloudApiManager:
     def list_all_ckpts(self, username=None, user_token=""):
         try:
             if self.auth_manger.enableAuth and not user_token:
+                return []
+
+            if not has_config():
                 return []
 
             params = {
@@ -241,7 +248,7 @@ class CloudApiManager:
         return raw_resp.json()['data']
 
     def list_roles(self, user_token=""):
-        if not self.auth_manger.enableAuth:
+        if not self.auth_manger.enableAuth or not has_config():
             return {
                 'roles': []
             }
